requirements.txt
telegram
google-generativeai